revamp export schema bug

+++ packages/cli/src/commands/export-schema.ts
@@ -37,32 +37,48 @@ export default class ExportSchema extends Command {
     }
   }
 
-  async getDataMigrations(client: any): Promise<Migrations> {
-    const { data } = await client.getMigrations();
-    return data.items;
+  async getDataMigrations(client: DefaultApi): Promise<Migrations> {
+    const { data } = await client.getMigrations(undefined, undefined, {});
+    return data.items as Migrations;
   }
 
   async exportFile(location: string, data: Migrations) {
+    const existing = fs.readdirSync("./migrations").map(v => +v.split("-")[0]);
+    let emptyCheck = true;
     data.forEach(file => {
+      if (!existing.includes(file.id)) {
+        emptyCheck = false;
+        this.log(
+          `${chalk.grey(
+            `#${file.id} ${new Date(file.createdAt).toISOString()}-${file.name}`
+          )}`
+        );
+        fs.writeFile(
+          `${location}/${file.id}-${new Date(
+            file.createdAt
+          ).toISOString()}.json`,
+          JSON.stringify(file, null, 2),
+          {
+            encoding: "utf8",
+            flag: "w",
+            mode: 0o666
+          },
+          err => {
+            if (err) return this.error(`${chalk.red(err)}`);
+          }
+        );
+      }
+    });
+    if (emptyCheck)
       this.log(
-        `${chalk.grey(
-          `#${file.id} ${new Date(file.createdAt).toISOString()}-${file.name}`
+        `${chalk.green(
+          "No newer migration, your migration is already up to date"
         )}`
       );
-
-      fs.writeFile(
-        `${location}/${file.id}-${new Date(file.createdAt).toISOString()}.json`,
-        JSON.stringify(file, null, 2),
-        {
-          encoding: "utf8",
-          flag: "w",
-          mode: 0o666
-        },
-        err => {
-          if (err) return this.error(`${chalk.red(err)}`);
-        }
-      );
-    });
+    else {
+      this.log(`${chalk.grey("\n\nExport file migrations")} ...`);
+      this.log(`${chalk.green("\nSuccess\n\n")}`);
+    }
   }
 
   async run() {
@@ -76,7 +92,5 @@ export default class ExportSchema extends Command {
     this.log(`${chalk.yellow("\n\nRunning process")} ...\n`);
     const data = await this.getDataMigrations(client);
     await this.exportFile(location, data);
-    this.log(`${chalk.grey("\n\nExport file migrations")} ...`);
-    this.log(`${chalk.green("\nSuccess\n\n")}`);
   }
 }
